feat: update coding standards and refactor GraphQL adapters
- Revamped `.editorconfig` with comprehensive C# coding conventions and preferences.
- Added `System.Text.RegularExpressions` usage in `AssemblySelector.cs` and simplified `RegexCache`.
- Introduced `CategoryGraphType` for GraphQL category representation.
- Implemented `GetAllCategoryGraphQLNetAdapter` and `GetAllCategoryHotChocolateAdapter` for asynchronous category queries.
- Enhanced `GetAllCategoryQuery` for better async handling.
- Updated static methods in `RequestResult{T}` to use the `new` modifier correctly.
- Corrected assembly formatting in `CommonServicesExtension.cs`.
- Improved parameter naming in `Player.razor.cs`.
- Cleaned up unnecessary package references in project files.
- Refactored `GraphQLAdapter` classes to align with updated coding standards.
