+++ .editorconfig
@@ -194,11 +194,11 @@ dotnet_naming_rule.interface_should_be_begins_with_i.style = begins_with_i
 
 dotnet_naming_symbols.interface.applicable_kinds = interface
 dotnet_naming_symbols.interface.applicable_accessibilities = public, internal, private, protected, protected_internal, private_protected
-dotnet_naming_symbols.interface.required_modifiers = 
+dotnet_naming_symbols.interface.required_modifiers =
 
 dotnet_naming_style.begins_with_i.required_prefix = I
-dotnet_naming_style.begins_with_i.required_suffix = 
-dotnet_naming_style.begins_with_i.word_separator = 
+dotnet_naming_style.begins_with_i.required_suffix =
+dotnet_naming_style.begins_with_i.word_separator =
 dotnet_naming_style.begins_with_i.capitalization = pascal_case
 
 # Enum naming rule
@@ -208,11 +208,11 @@ dotnet_naming_rule.enum_should_be_begins_with_e.style = begins_with_e
 
 dotnet_naming_symbols.enum.applicable_kinds = enum
 dotnet_naming_symbols.enum.applicable_accessibilities = public, internal, private, protected, protected_internal, private_protected
-dotnet_naming_symbols.enum.required_modifiers = 
+dotnet_naming_symbols.enum.required_modifiers =
 
 dotnet_naming_style.begins_with_e.required_prefix = E
-dotnet_naming_style.begins_with_e.required_suffix = 
-dotnet_naming_style.begins_with_e.word_separator = 
+dotnet_naming_style.begins_with_e.required_suffix =
+dotnet_naming_style.begins_with_e.word_separator =
 dotnet_naming_style.begins_with_e.capitalization = pascal_case
 
 # Constant fields naming rule
@@ -356,3 +356,6 @@ dotnet_style_require_accessibility_modifiers = none
 
 [*.{cs,vb}]
 dotnet_analyzer_diagnostic.category-Style.severity = none
+
+[*.cs]
+dotnet_diagnostic.SA1503.severity = none
