+++ src/Core/OnForkHub.Application/GraphQL/Mutations/Categories/CategoryGraphType.cs
@@ -1,0 +1,12 @@
+using GraphQL.Types;
+
+namespace OnForkHub.Application.GraphQL.Mutations.Categories;
+
+public class CategoryGraphType : ObjectGraphType<Category>
+{
+    public CategoryGraphType()
+    {
+        Field(x => x.Id);
+        Field(x => x.Name);
+    }
+}
