+++ test/Core/OnForkHub.Application.Test/OnForkHub.Application.Test.csproj
@@ -22,7 +22,6 @@
   </ItemGroup>
   <ItemGroup>
     <PackageReference Include="NSubstitute" />
-    <PackageReference Include="xunit" />
     <PackageReference Include="FluentAssertions" />
   </ItemGroup>
 </Project>
